requirements.txt
asgiref==3.5.2
coverage==6.5.0
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
psycopg2-binary==2.9.4
pytz==2022.4
sqlparse==0.4.3
tzdata==2022.4
